chore(fuelcalc): rewrite using ts
