--- app/fuelcalc/.babelrc
@@ -1,7 +1,0 @@
-{
-	"env": {
-		"test": {
-			"presets": [["preact-cli/babel", { "modules": "commonjs" }]]
-		}
-	}
-}
