--- app/fuelcalc/.eslintrc
@@ -1,1 +1,0 @@
-{ "root": true, "extends": "preact" }
