+++ app/fuelcalc/package.json
@@ -1,31 +1,22 @@
 {
 	"private": true,
-	"name": "fuelcalc",
-	"version": "0.0.0",
-	"license": "MIT",
+	"type": "module",
 	"scripts": {
-		"predeploy": "preact build --sw false --dest demo",
-		"serve": "sirv build --cors --single",
-		"dev": "preact watch",
-		"lint": "eslint src"
-	},
-	"eslintConfig": {
-		"extends": "preact",
-		"ignorePatterns": [
-			"build/"
-		]
-	},
-	"devDependencies": {
-		"eslint": "^8.56.0",
-		"eslint-config-preact": "^1.3.0",
-		"identity-obj-proxy": "^3.0.0",
-		"preact-cli": "^3.0.0-rc.6",
-		"preact-render-spy": "^1.2.1",
-		"sirv-cli": "^0.4.5"
+		"dev": "vite",
+		"build": "vite build",
+		"preview": "vite preview"
 	},
 	"dependencies": {
-		"preact": "^10.3.2",
-		"preact-render-to-string": "^5.1.4",
-		"preact-router": "^3.2.1"
+		"preact": "^10.13.1"
+	},
+	"devDependencies": {
+		"@preact/preset-vite": "^2.5.0",
+		"eslint": "^8.56.0",
+		"eslint-config-preact": "^1.3.0",
+		"typescript": "^5.3.3",
+		"vite": "^4.3.2"
+	},
+	"eslintConfig": {
+		"extends": "preact"
 	}
 }
