--- app/fuelcalc/size-plugin.json
@@ -1,1 +1,0 @@
-[{"timestamp":1705758406250,"files":[{"filename":"bundle.*****.esm.js","previous":7106,"size":7114,"diff":8},{"filename":"polyfills.*****.esm.js","previous":2319,"size":2322,"diff":3},{"filename":"polyfills.84370.js","previous":2419,"size":0,"diff":-2419},{"filename":"index.html","previous":1335,"size":1339,"diff":4},{"filename":"bundle.09f19.css","previous":578,"size":578,"diff":0},{"filename":"bundle.dc492.js","previous":7418,"size":0,"diff":-7418},{"filename":"bundle.e6999.js","previous":0,"size":7426,"diff":7426},{"filename":"polyfills.40b7a.js","previous":0,"size":2423,"diff":2423}]},{"timestamp":1705755888006,"files":[{"filename":"bundle.f4bd4.css","previous":577,"size":0,"diff":-577},{"filename":"bundle.*****.esm.js","previous":7104,"size":7106,"diff":2},{"filename":"polyfills.*****.esm.js","previous":2319,"size":2319,"diff":0},{"filename":"bundle.b9374.js","previous":7416,"size":0,"diff":-7416},{"filename":"polyfills.84370.js","previous":2419,"size":2419,"diff":0},{"filename":"index.html","previous":1335,"size":1335,"diff":0},{"filename":"bundle.09f19.css","previous":0,"size":578,"diff":578},{"filename":"bundle.dc492.js","previous":0,"size":7418,"diff":7418}]},{"timestamp":1705702816170,"files":[{"filename":"ssr-build/ssr-bundle.f4bd4.css","previous":636,"size":0,"diff":-636},{"filename":"ssr-build/ssr-bundle.js","previous":5431,"size":0,"diff":-5431},{"filename":"bundle.f4bd4.css","previous":0,"size":577,"diff":577},{"filename":"bundle.*****.esm.js","previous":0,"size":7104,"diff":7104},{"filename":"polyfills.*****.esm.js","previous":0,"size":2319,"diff":2319},{"filename":"bundle.b9374.js","previous":0,"size":7416,"diff":7416},{"filename":"polyfills.84370.js","previous":0,"size":2419,"diff":2419},{"filename":"index.html","previous":0,"size":1335,"diff":1335}]}]
