+++ app/fuelcalc/src/components/App.module.css
@@ -1,0 +1,3 @@
+.container {
+	max-width: 700px;
+}
