+++ app/fuelcalc/src/components/App.tsx
@@ -1,0 +1,12 @@
+import { Calc } from "./Form";
+import { Header } from "./Header";
+import style from "./App.module.css";
+
+export function App() {
+	return (
+		<div className={style.container}>
+			<Header />
+			<Calc />
+		</div>
+	);
+}
