+++ app/fuelcalc/src/components/Form.module.css
@@ -1,0 +1,30 @@
+.form {
+	background-color: rgba(200, 200, 200, 0.8);
+	display: flex;
+}
+.col {
+	width: 50%;
+	max-width: 100%;
+	padding: 15px;
+}
+.container {
+	display: flex;
+}
+.form-group label {
+	display: block;
+	margin-bottom: 0.25rem;
+}
+.form-control[type] {
+	display: block;
+	width: calc(100% - 2px - 1.4rem);
+	margin: 0;
+	padding: 0.375rem 0.7rem;
+}
+
+.formRow {
+	display: flex;
+	flex-direction: row;
+}
+
+.col {
+}
