+++ app/fuelcalc/src/components/Header.module.css
@@ -2,7 +2,7 @@
 	width: 100%;
 	height: 56px;
 	padding: 0;
-	background: #FF2800;
+	background: #ff2800;
 	box-shadow: 0 0 5px rgba(0, 0, 0, 0.5);
 	z-index: 50;
 }
@@ -14,7 +14,7 @@
 	font-size: 24px;
 	line-height: 56px;
 	font-weight: 400;
-	color: #FFF;
+	color: #fff;
 }
 
 .header nav {
@@ -29,17 +29,17 @@
 	padding: 0 15px;
 	min-width: 50px;
 	text-align: center;
-	background: rgba(255,255,255,0);
+	background: rgba(255, 255, 255, 0);
 	text-decoration: none;
-	color: #FFF;
+	color: #fff;
 	will-change: background-color;
 }
 
 .header nav a:hover,
 .header nav a:active {
-	background: rgba(0,0,0,0.2);
+	background: rgba(0, 0, 0, 0.2);
 }
 
 .header nav a.active {
-	background: rgba(0,0,0,0.4);
+	background: rgba(0, 0, 0, 0.4);
 }
