+++ app/fuelcalc/src/components/Header.tsx
@@ -1,0 +1,7 @@
+import style from "./Header.module.css";
+
+export const Header = () => (
+	<header class={style.header}>
+		<h1>Fuel Calculator</h1>
+	</header>
+);
