--- app/fuelcalc/src/components/calc/style.css
@@ -1,3 +1,0 @@
-form {
-    background-color: rgba(200,200,200, 0.8);
-}
