--- app/fuelcalc/src/components/header/index.js
@@ -1,10 +1,0 @@
-import { h } from "preact";
-import style from "./style.css";
-
-const Header = () => (
-	<header class={style.header}>
-		<h1>Fuel Calculator</h1>
-	</header>
-);
-
-export default Header;
