--- app/fuelcalc/src/components/home/index.js
@@ -1,11 +1,0 @@
-import { h } from "preact";
-import style from "./style";
-
-const Home = () => (
-	<div class={style.home}>
-		<h1>Home</h1>
-		<p>This is the Home component.</p>
-	</div>
-);
-
-export default Home;
