--- app/fuelcalc/src/components/home/style.css
@@ -1,5 +1,0 @@
-.home {
-	padding: 56px 20px;
-	min-height: 100%;
-	width: 100%;
-}
