--- app/fuelcalc/src/index.js
@@ -1,4 +1,0 @@
-import "./style/index.css";
-import App from "./components/app";
-
-export default App;
