+++ app/fuelcalc/src/index.tsx
@@ -1,0 +1,6 @@
+import { render } from "preact";
+
+import "./style.css";
+import { App } from "./components/App";
+
+render(<App />, document.getElementById("app"));
