--- app/fuelcalc/src/manifest.json
@@ -1,21 +1,0 @@
-{
-  "name": "fuelcalc",
-  "short_name": "fuelcalc",
-  "start_url": "/fuelcalc",
-  "display": "standalone",
-  "orientation": "portrait",
-  "background_color": "#fff",
-  "theme_color": "#FF2800",
-  "icons": [
-    {
-      "src": "/assets/icons/android-chrome-192x192.png",
-      "type": "image/png",
-      "sizes": "192x192"
-    },
-    {
-      "src": "/assets/icons/android-chrome-512x512.png",
-      "type": "image/png",
-      "sizes": "512x512"
-    }
-  ]
-}
