+++ app/fuelcalc/src/style.css
@@ -1,0 +1,37 @@
+:root {
+	font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+	line-height: 1.5;
+	font-weight: 400;
+
+	--color-text: #222;
+	--color-bg: #ffffff;
+
+	font-synthesis: none;
+	text-rendering: optimizeLegibility;
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
+	-webkit-text-size-adjust: 100%;
+}
+
+@media screen and (prefer-color-scheme: dark) {
+	:root {
+		--color-text: #ccc;
+		--color-bg: #1a1a1a;
+	}
+}
+
+body {
+	background-image: url(https://live.staticflickr.com/4844/45573295934_555c082cb2_k.jpg);
+	background-size: cover;
+	background-repeat: no-repeat;
+	background-position: center center;
+
+	margin: 0;
+	display: flex;
+	flex-direction: column;
+	justify-content: flex-start;
+	align-items: center;
+	min-height: 100vh;
+	color: var(--color-text);
+	background-color: var(--color-bg);
+}
