--- app/fuelcalc/src/style/index.css
@@ -1,27 +1,0 @@
-html, body {
-	height: 100%;
-	width: 100%;
-	padding: 0;
-	margin: 0;
-	background: #FAFAFA;
-	font-family: 'Helvetica Neue', arial, sans-serif;
-	font-weight: 400;
-	color: #444;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
-}
-
-* {
-	box-sizing: border-box;
-}
-
-#app {
-	height: 100%;
-}
-
-body {
-	background-image: url(https://live.staticflickr.com/4844/45573295934_555c082cb2_k.jpg);
-	background-size:     cover;
-	background-repeat:   no-repeat;
-	background-position: center center;
-}
