+++ app/fuelcalc/vite.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from 'vite';
+import preact from '@preact/preset-vite';
+
+// https://vitejs.dev/config/
+export default defineConfig({
+	plugins: [preact()],
+});
